Consolidate common utils in extensions
