+++ modules/extensions/src/terrain/shader-module.ts
@@ -2,7 +2,7 @@ import {project} from '@deck.gl/core';
 import type {_ShaderModule as ShaderModule} from '@deck.gl/core';
 
 import type {Texture2D} from '@luma.gl/core';
-import type {Bounds} from '../common/projection-utils';
+import type {Bounds} from '../utils/projection-utils';
 import type {TerrainCover} from './terrain-cover';
 
 /** Module parameters expected by the terrain shader module */
