+++ modules/extensions/src/utils/projection-utils.ts
